requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pymongo==4.3.3
python-dotenv==1.0.0
httpx==0.24.0
pytest==7.3.1
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
